PythonAPI/requirements.txt
Flask==3.1.2
flask_cors==6.0.2
langchain_classic==1.0.0
langchain_community==0.4.1
langchain_huggingface==1.2.0
langchain_mistralai==1.1.1
pandas==2.3.3

pypdf==6.4.1
sentence-transformers==5.2.0
faiss-cpu
python-dotenv